requirements.txt
playwright
requests
beautifulsoup4
pandas
pyarrow
tqdm
tenacity
python-slugify